+++ sklearn_genetic/__init__.py
@@ -1,6 +1,10 @@
 from .genetic_search import GASearchCV
 from .plots import plot_fitness_evolution, plot_search_space
-from .callbacks import ThresholdStopping, ConsecutiveStopping, DeltaThreshold
+from .callbacks import (
+    ThresholdStopping,
+    ConsecutiveStopping,
+    DeltaThreshold,
+)
 from ._version import __version__
 
 __all__ = [
